Move existing action creator to new directory
